Rename load-files to find-files
