+++ packages/chookscord/src/tools/find-files.ts
